resolve conflict, tiles setup, 양식 통일, 작업공간 구분

+++ src/main/java/com/winebouti/vo/ReplyVO.java
@@ -6,9 +6,9 @@ import lombok.ToString;
 
 @Getter @Setter @ToString
 public class ReplyVO {
-	private int productId;
-    private int memberId;
-    private int purchaseId;
+	private long productId;
+    private long memberId;
+    private long purchaseId;
     private String content;
     private int star;
     private boolean isOpen;
