+++ src/main/java/com/winebouti/vo/ReviewVO.java
@@ -7,12 +7,12 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ReviewVO {
-	
-	    private int review_id;
-	    private int product_id;
-	    private String user_name;
-	    private String content;
-	    private int rating;
-	    private String created_at;
+
+	private long review_id;
+	private long product_id;
+	private String user_name;
+	private String content;
+	private int rating;
+	private String created_at;
 
 }
